+++ src/main.rs
@@ -1,9 +1,10 @@
-use pulsar_web::{Pulse};
+use pulsar_web::Pulse;
 
 fn main() {
     let mut server = Pulse::new(3000);
-    server.get("/", move |_| {
-        ("hello, world!".to_string(), 200)
+    server.get("/", |req, res| {
+        res.body("hello world!");
+        res.clone()
     });
     server.launch();
 }