+++ changelog.md
@@ -1,5 +1,12 @@
 # Changelog
 
+### 0.0.2
+
+- added `req` and `res` arguments for request methods
+- added `POST`, `PUT`, `PATCH`, `DELETE`
+- updated method syntax
+- 
+
 ### 0.0.1
 
 - added `GET` method handler